Refactor a lot

+++ src/main/java/com/myname/mymodid/Rendering/PlayerPosition.java
@@ -2,16 +2,16 @@ package com.myname.mymodid.Rendering;
 
 public class PlayerPosition {
 
-    public PlayerPosition(double x, double y, double z, long time) {
+    public PlayerPosition(int x, int y, int z, long time) {
         this.x = x;
         this.y = y;
         this.z = z;
         this.time = time;
     }
 
-    public double time;
-    public double x;
-    public double y;
-    public double z;
+    public long time;
+    public int x;
+    public int y;
+    public int z;
 
 }
